backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
celery==5.3.6
redis==5.0.1
reportlab==4.0.9
lxml==5.1.0
cryptography==42.0.2
requests==2.31.0
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2024.1
whitenoise==6.6.0
gunicorn==21.2.0
django-filter==24.1
drf-spectacular==0.27.1
